renaming parcel_carrier to carrier and removing parcel word from service and repository name
