+++ internal/app/service/service.go
@@ -7,12 +7,12 @@ import (
 )
 
 //go:generate mockgen -source service.go -destination ./mocks/mock_service.go -package mocks
-// CarrierParcelAcceptRepository to update the carrier request table
-type CarrierParcelAcceptRepository interface {
+// CarrierAcceptRepository to update the carrier request table
+type CarrierAcceptRepository interface {
 	UpdateCarrierRequest(ctx context.Context, parcel model.CarrierRequest, acceptStatus int, rejectStatus int, parcelStatus int, sourceTime time.Time) error
 }
 
 // UserService to fetch user by PhoneNumber and Password
-type CarrierParcelAcceptService interface {
+type CarrierAcceptService interface {
 	AssignCarrierToParcel(ctx context.Context, parcel model.CarrierRequest) error
 }
